Add support for databricks (Close #77)

+++ macros/utils/post_ci_cleanup.sql
@@ -1,26 +1,37 @@
 {# Destructive macro. Use with care! #}
 
-{% macro post_ci_cleanup(schema_pattern=target.schema~'%') %}
+{% macro post_ci_cleanup(schema_pattern=target.schema) %}
   
   {# Get all schemas with the target.schema prefix #}
-  {% set get_tables_sql = dbt_utils.get_tables_by_pattern_sql(schema_pattern,table_pattern='%') %}
-  {# If no schemas exists then get_tables_sql is empty string so don't query #}
-  {% set results = [] if get_tables_sql.isspace() else run_query(get_tables_sql) %}
-  {% set schemas = results|map(attribute='table_schema')|unique|list %}
+  {% set schemas = snowplow_utils.get_schemas_by_pattern(schema_pattern,table_pattern='%') %}
 
   {% if schemas|length %}
 
-    {# Generate sql to drop all identified schemas #}
-    {% set drop_schema_sql -%}
-
+    {%- if target.type == 'databricks' -%}
+      {# Generate sql to drop all identified schemas #}
       {% for schema in schemas -%}
-        DROP SCHEMA IF EXISTS {{schema}} CASCADE; 
+        {%- set drop_schema_sql -%}
+          DROP SCHEMA IF EXISTS {{schema}} CASCADE; 
+        {%- endset -%}
+
+        {% do run_query(drop_schema_sql) %}
+
       {% endfor %}
+    
+    {%- else -%}
+      {# Generate sql to drop all identified schemas #}
+      {% set drop_schema_sql -%}
 
-    {%- endset %}
+        {% for schema in schemas -%}
+          DROP SCHEMA IF EXISTS {{schema}} CASCADE; 
+        {% endfor %}
 
-    {# Drop schemas #}
-    {% do run_query(drop_schema_sql) %}
+      {%- endset %}
+
+      {# Drop schemas #}
+      {% do run_query(drop_schema_sql) %}
+    
+    {%- endif -%}
 
   {% endif %}
